refactor(orb-ui): refactor interfaces for sink backends and improve sink promises resolution
